refactor: Improves SCRAM code reuse
